+++ core/action-queue/tests/common/mod.rs
@@ -1,9 +1,10 @@
 #![allow(dead_code)]
 use std::sync::Arc;
 
-use proton_action_queue::action::{Action, Error, Factory, WriterGuardError};
+use proton_action_queue::action::{Action, Factory};
 use proton_action_queue::network::WaitForOnline;
 use proton_action_queue::queue::Queue;
+pub use proton_action_queue::tests::common::DefaultError;
 use stash::exports::SqliteError;
 use stash::params;
 use stash::stash::{Bond, Stash, StashError, Tether};
@@ -148,38 +149,5 @@ impl TestWriteExtension for Bond<'_> {
         self.execute("DELETE FROM ext WHERE key=?", params![key.to_owned()])
             .await?;
         Ok(())
-    }
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum DefaultError {
-    #[error("Network Failure")]
-    NetworkFailure,
-    #[error("API Failure")]
-    APIFailure,
-    #[error("{0}")]
-    Other(anyhow::Error),
-    #[error("{0}")]
-    DB(#[from] StashError),
-    #[error("Writer Guard Expired")]
-    WriterGuardExpired,
-}
-
-impl From<WriterGuardError> for DefaultError {
-    fn from(value: WriterGuardError) -> Self {
-        match value {
-            WriterGuardError::Expired => Self::WriterGuardExpired,
-            WriterGuardError::Stash(e) => Self::DB(e),
-        }
-    }
-}
-
-impl Error for DefaultError {
-    fn is_network_failure(&self) -> bool {
-        matches!(self, DefaultError::NetworkFailure)
-    }
-
-    fn is_writer_guard_expired(&self) -> bool {
-        matches!(self, DefaultError::WriterGuardExpired)
     }
 }